Add build github actions workflow
