--- .dockerignore
@@ -1,2 +1,0 @@
-lambda-registrator
-fixtures
