+++ .github/CODEOWNERS
@@ -1,0 +1,6 @@
+# default PR reviews to the team
+* @hashicorp/consul-ecs
+
+# release configuration
+/.release/                              @hashicorp/release-engineering
+/.github/workflows/build.yml            @hashicorp/release-engineering
