+++ consul-lambda-registrator/.dockerignore
@@ -1,0 +1,2 @@
+consul-lambda-registrator
+fixtures
