+++ consul-lambda-registrator/.go-version
@@ -1,0 +1,1 @@
+1.18.2
