+++ consul-lambda-registrator/fixtures/create_function.json
@@ -1,0 +1,8 @@
+{
+    "detail": {
+        "eventName": "CreateFunction20150331",
+        "responseElements": {
+            "functionArn": "arn:aws:lambda:us-east-1:111111111111:function:lambda-1234"
+        }
+    }
+}
