--- lambda-registrator/fixtures/unsupported.json
@@ -1,8 +1,0 @@
-{
-    "detail": {
-        "eventName": "Unsupported",
-        "responseElements": {
-            "functionArn": "arn:aws:lambda:us-east-1:111111111111:function:lambda-1234"
-        }
-    }
-}
