--- lambda-registrator/fixtures/untag_resource.json
@@ -1,8 +1,0 @@
-{
-    "detail": {
-        "eventName": "UntagResource20170331v2",
-        "requestParameters": {
-            "resource": "arn:aws:lambda:us-east-1:111111111111:function:lambda-1234"
-        }
-    }
-}
